fix error preventing script from running

+++ Utility/ApiToolLogging.py
@@ -50,7 +50,7 @@ class ApiToolLog:
         self.limitFileSize(self.placePath)
 
     def limitFileSize(self, file, maxFileSizeInMb=5):
-        if (os.path.getsize(file) / (1024 * 1024)) > maxFileSizeInMb:
+        if os.path.exists(file) and (os.path.getsize(file) / (1024 * 1024)) > maxFileSizeInMb:
             with open(file, "w"):
                 pass
